Home Assistant is working again (async)
Home Assistant is working again. Unfortunately, the code is async, so it needs the rest of the app fixed too. It should fix issue #158
